chore(infrastructure): createInternalFunctionAsync moved to separate file
